chore: renaming folders
