+++ projects/design-angular-kit/schematics/ng-add/rules/setup-project/add-design-angular-kit.ts
