+++ projects/design-angular-kit/schematics/ng-add/rules/setup-project/add-http-client.ts
