+++ projects/design-angular-kit/schematics/ng-add/rules/setup-project/index.ts
